+++ src/effect_compressor.h
@@ -6,20 +6,23 @@
 #pragma once
 
 #include <atomic>
+#include <cmath>
+#include <cstdint>
 
 #include "compressor.h"
 
 class AudioEffectCompressor
 {
 public:
-	static const unsigned CompressorRatioInf = 31;
+	static const int CompressorRatioInf = 31;
 
-	AudioEffectCompressor(float samplerate):
+	AudioEffectCompressor(float samplerate) :
 	bypass{},
 	samplerate{samplerate},
 	compL{samplerate},
 	compR{samplerate}
-	{}
+	{
+	}
 
 	void setPreGain_dB(float gain)
 	{
@@ -71,13 +74,13 @@ public:
 		compR.resetStates();
 	}
 
-	void process(float32_t* blockL, float32_t* blockR, uint16_t len)
+	void process(float *blockL, float *blockR, int len)
 	{
 		if (bypass) return;
 
-		compL.doCompression (blockL, len);
-		compR.doCompression (blockR, len);
-    	}
+		compL.doCompression(blockL, static_cast<uint16_t>(len));
+		compR.doCompression(blockR, static_cast<uint16_t>(len));
+	}
 
 	std::atomic<bool> bypass;
